+++ lib/rubocop/cop/layout/space_after_method_name.rb
@@ -9,9 +9,11 @@ module RuboCop
       #
       #   # bad
       #   def func (x) ... end
+      #   def method= (y) ... end
       #
       #   # good
       #   def func(x) ... end
+      #   def method=(y) ... end
       class SpaceAfterMethodName < Cop
         MSG = 'Do not put a space between a method name and the opening ' \
               'parenthesis.'.freeze