+++ lib/rubocop/cop/style/comment_annotation.rb
@@ -5,6 +5,31 @@ module RuboCop
     module Style
       # This cop checks that comment annotation keywords are written according
       # to guidelines.
+      #
+      # @example
+      #   # bad
+      #   # TODO make better
+      #
+      #   # good
+      #   # TODO: make better
+      #
+      #   # bad
+      #   # TODO:make better
+      #
+      #   # good
+      #   # TODO: make better
+      #
+      #   # bad
+      #   # fixme: does not work
+      #
+      #   # good
+      #   # FIXME: does not work
+      #
+      #   # bad
+      #   # Optimize does not work
+      #
+      #   # good
+      #   # OPTIMIZE: does not work
       class CommentAnnotation < Cop
         include AnnotationComment