+++ lib/rubocop/cop/style/def_with_parentheses.rb
@@ -6,6 +6,33 @@ module RuboCop
       # This cop checks for parentheses in the definition of a method,
       # that does not take any arguments. Both instance and
       # class/singleton methods are checked.
+      #
+      # @example
+      #
+      #   # bad
+      #   def foo()
+      #     # does a thing
+      #   end
+      #
+      #   # good
+      #   def foo
+      #     # does a thing
+      #   end
+      #
+      #   # also good
+      #   def foo() does_a_thing end
+      #
+      # @example
+      #
+      #   # bad
+      #   def Baz.foo()
+      #     # does a thing
+      #   end
+      #
+      #   # good
+      #   def Baz.foo
+      #     # does a thing
+      #   end
       class DefWithParentheses < Cop
         MSG = "Omit the parentheses in defs when the method doesn't accept " \
               'any arguments.'.freeze