+++ spec/rubocop/cop/lint/redundant_with_index_spec.rb
@@ -19,6 +19,13 @@ describe RuboCop::Cop::Lint::RedundantWithIndex do
     RUBY
   end
 
+  it 'registers an offense when using `ary.each.with_index(1) { |v| v }`' do
+    expect_offense(<<-RUBY.strip_indent)
+      ary.each.with_index(1) { |v| v }
+               ^^^^^^^^^^^^^ Remove redundant `with_index`.
+    RUBY
+  end
+
   it 'registers an offense when using `ary.each_with_object([]).with_index ' \
      '{ |v| v }`' do
     expect_offense(<<-RUBY.strip_indent)
@@ -35,6 +42,12 @@ describe RuboCop::Cop::Lint::RedundantWithIndex do
 
   it 'autocorrects to ary.each from ary.each.with_index' do
     new_source = autocorrect_source('ary.each.with_index { |v| v }')
+
+    expect(new_source).to eq 'ary.each { |v| v }'
+  end
+
+  it 'autocorrects to ary.each from ary.each.with_index(1)' do
+    new_source = autocorrect_source('ary.each.with_index(1) { |v| v }')
 
     expect(new_source).to eq 'ary.each { |v| v }'
   end