+++ spec/rubocop/cop/style/send_spec.rb
@@ -32,8 +32,6 @@ describe RuboCop::Cop::Style::Send do
   end
 
   context 'with __send__' do
-    after { expect(cop.offenses.empty?).to be(true) }
-
     context 'and with a receiver' do
       it 'does not register an offense for an invocation with args' do
         expect_no_offenses('Object.__send__(:inspect)')
@@ -56,8 +54,6 @@ describe RuboCop::Cop::Style::Send do
   end
 
   context 'with public_send' do
-    after { expect(cop.offenses.empty?).to be(true) }
-
     context 'and with a receiver' do
       it 'does not register an offense for an invocation with args' do
         expect_no_offenses('Object.public_send(:inspect)')