+++ spec/rubocop/cop/style/trailing_underscore_variable_spec.rb
@@ -6,159 +6,127 @@ describe RuboCop::Cop::Style::TrailingUnderscoreVariable do
   shared_examples 'common functionality' do
     it 'registers an offense when the last variable of parallel assignment ' \
        'is an underscore' do
-      inspect_source('a, b, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, b, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, b, _ = foo()
+              ^^ Do not use trailing `_`s in parallel assignment. Prefer `a, b, = foo()`.
+      RUBY
     end
 
     it 'registers an offense when multiple underscores are used '\
        'as the last variables of parallel assignment ' do
-      inspect_source('a, _, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, _, _ = foo()
+           ^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'registers an offense for splat underscore as the last variable' do
-      inspect_source('a, *_ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, *_ = foo()
+           ^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'registers an offense when underscore is the second to last variable ' \
        'and blank is the last variable' do
-      inspect_source('a, _, = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, _, = foo()
+           ^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'registers an offense when underscore is the only variable ' \
        'in parallel assignment' do
-      inspect_source('_, = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        _, = foo()
+        ^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `foo()`.
+      RUBY
     end
 
     it 'registers an offense for an underscore as the last param ' \
        'when there is also an underscore as the first param' do
-      inspect_source('_, b, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `_, b, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        _, b, _ = foo()
+              ^^ Do not use trailing `_`s in parallel assignment. Prefer `_, b, = foo()`.
+      RUBY
     end
 
     it 'does not register an offense when there are no underscores' do
-      inspect_source('a, b, c = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, b, c = foo()')
     end
 
     it 'does not register an offense for underscores at the beginning' do
-      inspect_source('_, a, b = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('_, a, b = foo()')
     end
 
     it 'does not register an offense for an underscore preceded by a ' \
        'splat variable anywhere in the argument chain' do
-      inspect_source('*a, b, _ = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('*a, b, _ = foo()')
     end
 
     it 'does not register an offense for an underscore preceded by a ' \
        'splat variable' do
-      inspect_source('a, *b, _ = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _ = foo()')
     end
 
     it 'does not register an offense for an underscore preceded by a ' \
        'splat variable and another underscore' do
-      inspect_source('_, *b, _ = *foo')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('_, *b, _ = *foo')
     end
 
     it 'does not register an offense for multiple underscores preceded by a ' \
        'splat variable' do
-      inspect_source('a, *b, _, _ = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _, _ = foo()')
     end
 
     it 'does not register an offense for multiple named underscores ' \
        'preceded by a splat variable' do
-      inspect_source('a, *b, _c, _d = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _c, _d = foo()')
     end
 
     it 'registers an offense for multiple underscore variables preceded by ' \
        'a splat underscore variable' do
-      inspect_source('a, *_, _, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, *_, _, _ = foo()
+           ^^^^^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'registers an offense for nested assignments with trailing ' \
       'underscores' do
-      inspect_source('a, (b, _) = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, (b,) = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, (b, _) = foo()
+              ^^ Do not use trailing `_`s in parallel assignment. Prefer `a, (b,) = foo()`.
+      RUBY
     end
 
     it 'registers an offense for complex nested assignments with trailing ' \
       'underscores' do
-      inspect_source('a, (_, (b, _), *_) = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, (_, (b,), *_) = foo()`.',
-                'Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, (_, (b, _),) = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, (_, (b, _), *_) = foo()
+                  ^^ Do not use trailing `_`s in parallel assignment. Prefer `a, (_, (b,), *_) = foo()`.
+                      ^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, (_, (b, _),) = foo()`.
+      RUBY
     end
 
     it 'does not register an offense for a named underscore variable ' \
        'preceded by a splat variable' do
-      inspect_source('a, *b, _c = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _c = foo()')
     end
 
     it 'does not register an offense for a named variable preceded by a ' \
        'names splat underscore variable' do
-      inspect_source('a, *b, _c = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _c = foo()')
     end
 
     it 'does not register an offense for nested assignments without ' \
       'trailing underscores' do
-      inspect_source('a, (_, b) = foo()')
-
-      expect(cop.messages).to be_empty
+      expect_no_offenses('a, (_, b) = foo()')
     end
 
     it 'does not register an offense for complex nested assignments without ' \
       'trailing underscores' do
-      inspect_source('a, (_, (b,), c, (d, e),) = foo()')
-
-      expect(cop.messages).to be_empty
+      expect_no_offenses('a, (_, (b,), c, (d, e),) = foo()')
     end
 
     describe 'autocorrect' do
@@ -245,104 +213,93 @@ describe RuboCop::Cop::Style::TrailingUnderscoreVariable do
   end
 
   context 'configured to allow named underscore variables' do
-    include_examples 'common functionality'
-
     let(:config) do
-      RuboCop::Config.new('Style/TrailingUnderscoreVariable' => {
-                            'Enabled' => true,
-                            'AllowNamedUnderscoreVariables' => true
-                          })
+      RuboCop::Config.new(
+        'Style/TrailingUnderscoreVariable' => {
+          'Enabled' => true,
+          'AllowNamedUnderscoreVariables' => true
+        }
+      )
     end
 
-    it 'does not register an offense for an underscore variable preceded ' \
-       'by a named splat underscore variable' do
-      inspect_source('a, *_b, _ = foo()')
-
-      expect(cop.messages.empty?).to be(true)
-    end
+    include_examples 'common functionality'
 
     it 'does not register an offense for named variables ' \
        'that start with an underscore' do
-      inspect_source('a, b, _c = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, b, _c = foo()')
     end
 
     it 'does not register an offense for a named splat underscore ' \
        'as the last variable' do
-      inspect_source('a, *_b = foo()')
+      expect_no_offenses('a, *_b = foo()')
+    end
 
-      expect(cop.messages.empty?).to be(true)
+    it 'does not register an offense for an underscore variable preceded ' \
+       'by a named splat underscore variable' do
+      expect_no_offenses('a, *_b, _ = foo()')
     end
 
     it 'does not register an offense for multiple underscore variables ' \
        'preceded by a named splat underscore variable' do
-      inspect_source('a, *_b, _, _ = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *_b, _, _ = foo()')
     end
   end
 
   context 'configured to not allow named underscore variables' do
-    include_examples 'common functionality'
-
     let(:config) do
-      RuboCop::Config.new('Style/TrailingUnderscoreVariable' => {
-                            'Enabled' => true,
-                            'AllowNamedUnderscoreVariables' => false
-                          })
+      RuboCop::Config.new(
+        'Style/TrailingUnderscoreVariable' => {
+          'Enabled' => true,
+          'AllowNamedUnderscoreVariables' => false
+        }
+      )
     end
+
+    include_examples 'common functionality'
 
     it 'registers an offense for named variables ' \
        'that start with an underscore' do
-      inspect_source('a, b, _c = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, b, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, b, _c = foo()
+              ^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, b, = foo()`.
+      RUBY
     end
 
     it 'registers an offense for a named splat underscore ' \
        'as the last variable' do
-      inspect_source('a, *_b = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, *_b = foo()
+           ^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'does not register an offense for a named underscore preceded by a ' \
        'splat variable' do
-      inspect_source('a, *b, _c = foo()')
-
-      expect(cop.messages.empty?).to be(true)
+      expect_no_offenses('a, *b, _c = foo()')
     end
 
     it 'registers an offense for an underscore variable preceded ' \
        'by a named splat underscore variable' do
-      inspect_source('a, *_b, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, *_b, _ = foo()
+           ^^^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     it 'registers an offense for an underscore preceded by ' \
        'a named splat underscore' do
-      inspect_source('a, b, *_c, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, b, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, b, *_c, _ = foo()
+              ^^^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, b, = foo()`.
+      RUBY
     end
 
     it 'registers an offense for multiple underscore variables ' \
        'preceded by a named splat underscore variable' do
-      inspect_source('a, *_b, _, _ = foo()')
-
-      expect(cop.messages)
-        .to eq(['Do not use trailing `_`s in parallel assignment. ' \
-                'Prefer `a, = foo()`.'])
+      expect_offense(<<-RUBY.strip_indent)
+        a, *_b, _, _ = foo()
+           ^^^^^^^^^^ Do not use trailing `_`s in parallel assignment. Prefer `a, = foo()`.
+      RUBY
     end
 
     context 'autocorrect' do